requirements.txt
spotipy==2.23.0
typer[all]==0.9.0
python-dotenv==1.0.0
rich==13.7.0